+++ src/utils/tests/ContextTest.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://doc.hyperf.io
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf-cloud/hyperf/blob/master/LICENSE
+ */
+
+namespace HyperfTest\Utils;
+
+use Hyperf\Utils\Context;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @internal
+ * @coversNothing
+ */
+class ContextTest extends TestCase
+{
+    public function testOverride()
+    {
+        Context::set('override.id', 1);
+
+        $this->assertSame(2, Context::override('override.id', function ($id) {
+            return $id + 1;
+        }));
+
+        $this->assertSame(2, Context::get('override.id'));
+    }
+}